Empty HashMap should work correctly without an insert. (#3)
At the moment, the operations `contains_key`, `get` and `remove` does
not work on the empty HashMap. Because, it tries to calculate the hash
of key using zero (number of buckets) as a modulo.

This commit fixes the issue by adding early return to these methods.

Additionally, adds a test case for the fix.